fix(persistence): LIMIT/OFFSET を bigint 化し gosec G115(int→int32) を解消
sqlc が ::int で int32 パラメータを生成し int→int32 の縮小変換が gosec で HIGH 指摘。
クエリを ::bigint に変え int64（安全な拡大変換）にする。

+++ internal/infrastructure/persistence/postgres/character_repo.go
@@ -110,8 +110,8 @@ func (r *CharacterRepository) List(ctx context.Context, p domain.ListParams) ([]
 	}
 	rows, err := r.q.ListCharacters(ctx, sqlc.ListCharactersParams{
 		Ids: p.IDs, // nil なら $1 IS NULL で全件
-		Off: int32(p.Offset),
-		Lim: int32(p.Limit), // 0 なら NULLIF で LIMIT 無し
+		Off: int64(p.Offset),
+		Lim: int64(p.Limit), // 0 なら NULLIF で LIMIT 無し
 	})
 	if err != nil {
 		return nil, fmt.Errorf("postgres: list characters: %w", err)
